requirements.txt
python-dotenv
djangorestframework
django
black